You can now drop your current functional hat by pressing Q

+++ src/Hat.h
@@ -7,6 +7,6 @@ struct Hat {
   String128 name;
   String128 effect;
   uint8 id;
-  uint8 hatType;
+  int hatType;
   uint32 eid;
 };
